+++ packages/node/package.json
@@ -58,16 +58,6 @@
         "default": "./build/orchestrion/import-hook.mjs"
       }
     },
-    "./orchestrion/config": {
-      "import": {
-        "types": "./build/types/orchestrion/config.d.ts",
-        "default": "./build/esm/orchestrion/config.js"
-      },
-      "require": {
-        "types": "./build/types/orchestrion/config.d.ts",
-        "default": "./build/cjs/orchestrion/config.js"
-      }
-    },
     "./orchestrion/vite": {
       "import": {
         "types": "./build/types/orchestrion/bundler/vite.d.ts",
@@ -86,9 +76,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@apm-js-collab/code-transformer": "^0.13.0",
-    "@apm-js-collab/code-transformer-bundler-plugins": "^0.3.0",
-    "@apm-js-collab/tracing-hooks": "^0.8.0",
     "@opentelemetry/api": "^1.9.1",
     "@opentelemetry/core": "^2.6.1",
     "@opentelemetry/instrumentation": "^0.214.0",
@@ -101,16 +88,7 @@
     "import-in-the-middle": "^3.0.0"
   },
   "devDependencies": {
-    "@types/node": "^18.19.1",
-    "vite": "^5.0.0"
-  },
-  "peerDependencies": {
-    "vite": "^3.0.0 || ^4.0.0 || ^5.0.0 || ^6.0.0"
-  },
-  "peerDependenciesMeta": {
-    "vite": {
-      "optional": true
-    }
+    "@types/node": "^18.19.1"
   },
   "scripts": {
     "build": "run-p build:transpile build:types",
