+++ packages/node/src/orchestrion/index.ts
@@ -1,3 +1,3 @@
 export { _experimentalSetupOrchestrion } from './setup';
 export type { ExperimentalSetupOrchestrionOptions } from './setup';
-export { mysqlChannelIntegration } from '../integrations/tracing-channel/mysql';
+export { mysqlChannelIntegration } from '@sentry/server-utils/orchestrion';
