+++ packages/node/tsconfig.types.json
@@ -5,10 +5,6 @@
     "declaration": true,
     "declarationMap": true,
     "emitDeclarationOnly": true,
-    "outDir": "build/types",
-    // Required so Node16 module resolution can disambiguate package self-references
-    // (`@sentry/node/orchestrion/config` from inside this package) against the
-    // package's `.` export. Without this tsc reports TS2209.
-    "rootDir": "src"
+    "outDir": "build/types"
   }
 }
