+++ packages/server-utils/package.json
@@ -26,12 +26,63 @@
         "types": "./build/types/index.d.ts",
         "default": "./build/cjs/index.js"
       }
+    },
+    "./orchestrion": {
+      "import": {
+        "types": "./build/types/orchestrion/index.d.ts",
+        "default": "./build/esm/orchestrion/index.js"
+      },
+      "require": {
+        "types": "./build/types/orchestrion/index.d.ts",
+        "default": "./build/cjs/orchestrion/index.js"
+      }
+    },
+    "./orchestrion/config": {
+      "import": {
+        "types": "./build/types/orchestrion/config.d.ts",
+        "default": "./build/esm/orchestrion/config.js"
+      },
+      "require": {
+        "types": "./build/types/orchestrion/config.d.ts",
+        "default": "./build/cjs/orchestrion/config.js"
+      }
+    },
+    "./orchestrion/vite": {
+      "types": "./build/types/orchestrion/bundler/vite.d.ts",
+      "import": {
+        "default": "./build/esm/orchestrion/bundler/vite.js"
+      }
+    },
+    "./orchestrion/import-hook": {
+      "import": {
+        "default": "./build/orchestrion/import-hook.mjs"
+      }
     }
   },
   "typesVersions": {
     "<5.0": {
       "build/types/index.d.ts": [
         "build/types-ts3.8/index.d.ts"
+      ],
+      "orchestrion": [
+        "build/types-ts3.8/orchestrion/index.d.ts"
+      ],
+      "orchestrion/config": [
+        "build/types-ts3.8/orchestrion/config.d.ts"
+      ],
+      "orchestrion/vite": [
+        "build/types-ts3.8/orchestrion/bundler/vite.d.ts"
+      ]
+    },
+    "*": {
+      "orchestrion": [
+        "build/types/orchestrion/index.d.ts"
+      ],
+      "orchestrion/config": [
+        "build/types/orchestrion/config.d.ts"
+      ],
+      "orchestrion/vite": [
+        "build/types/orchestrion/bundler/vite.d.ts"
       ]
     }
   },
@@ -39,8 +90,24 @@
     "access": "public"
   },
   "dependencies": {
+    "@apm-js-collab/code-transformer-bundler-plugins": "^0.3.0",
+    "@apm-js-collab/code-transformer": "^0.13.0",
+    "@apm-js-collab/tracing-hooks": "^0.8.0",
     "@sentry/conventions": "^0.12.0",
-    "@sentry/core": "10.58.0"
+    "@sentry/core": "10.58.0",
+    "magic-string": "~0.30.0"
+  },
+  "devDependencies": {
+    "@types/node": "^18.19.1",
+    "vite": "^5.0.0"
+  },
+  "peerDependencies": {
+    "vite": "^3.0.0 || ^4.0.0 || ^5.0.0 || ^6.0.0"
+  },
+  "peerDependenciesMeta": {
+    "vite": {
+      "optional": true
+    }
   },
   "scripts": {
     "build": "run-p build:transpile build:types",
@@ -65,5 +132,18 @@
   "volta": {
     "extends": "../../package.json"
   },
-  "sideEffects": false
+  "sideEffects": false,
+  "nx": {
+    "targets": {
+      "build:transpile": {
+        "outputs": [
+          "{projectRoot}/build/esm",
+          "{projectRoot}/build/cjs",
+          "{projectRoot}/build/npm/esm",
+          "{projectRoot}/build/npm/cjs",
+          "{projectRoot}/build/orchestrion"
+        ]
+      }
+    }
+  }
 }
